Add static analysis rule to detect unsecure random number usage (#605)
* Add static analysis rule to detect unsecure random number usage

Implemented a new static code analysis rule that checks for the usage of unsecure random number generation in the context of cryptographic functions. This includes identifying calls to random:createIntInRange() for initialization vectors in AES-GCM and AES-ECB functions. Additionally, updated related files and tests to support this new rule.

* Refactor file paths in static analysis rules to be relative

Updated the file paths in rule<x>.json to use relative paths instead of absolute paths. This change improves portability and ensures that the rules can be used in different environments without modification.